requirements.txt
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
googleapis-common-protos==1.65.0
idna==3.8
importlib_metadata==8.4.0
iniconfig==2.0.0
mypy==1.11.2
mypy-extensions==1.0.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==4.25.4
pycparser==2.22
PyGithub==2.4.0
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
zipp==3.20.1
